# backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CMS Admin API (구현 사양)",
    "version": "0.1.0",
    "description": "백엔드(Go·Gin·헥사고날)가 실제로 구현/제공하는 관리자(admin) API 사양이다. FE 대시보드(refine)와 공유하기 위한 문서이며, docs/swagger.json(FE 요청 계약)·docs/api.md·docs/admin-auth.md 의 계약을 그대로 만족한다.\n\n## 네임스페이스\n- 관리자 API: `/admin/api/v1/...` (본 문서)\n- 사용자(public) API: `/api/v1/...` (snake_case, 별도)\n\n## refine simple-rest 규약\n- 목록: `_start`(포함)/`_end`(미포함) 페이지네이션, 응답 배열 + `X-Total-Count` 헤더.\n- 정렬: `_sort`(콤마)/`_order`(asc|desc, 콤마, 1:1).\n- 필터: `{field}_like`(부분 일치), `{field}`(정확 일치) 등. 화이트리스트 필드만 적용.\n- 필드 네이밍 camelCase. 수정은 `PATCH`(부분), 삭제는 200.\n- 에러 바디: `{ \"message\": \"...\" }`.\n\n## 인증/인가\n- JWT Bearer. 로그인/리프레시 외 모든 엔드포인트에 `Authorization: Bearer <accessToken>` 필요.\n- 권한 부족 403, 토큰 누락·만료·무효 401. 권한은 서버에서 강제(`resource:action`).\n- 역할: superadmin(우회)/editor(list·show·create·edit + comments:delete)/viewer(list·show).\n\n## 개발용 기본 계정 (admin_users 가 비어있을 때 자동 시드)\n- admin@example.com (superadmin) / editor@example.com (editor) / viewer@example.com (viewer)\n- 비밀번호: env `ADMIN_SEED_PASSWORD`(기본 `secret`). 운영 전 반드시 교체."
  },
  "servers": [
    { "url": "http://localhost:9000/admin/api/v1", "description": "로컬 개발 서버 (관리자 API)" },
    { "url": "http://localhost:8080/admin/api/v1", "description": "백엔드 기본 포트(HTTP_PORT 기본값 8080)" }
  ],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/posts": {
      "get": {
        "tags": ["posts"],
        "summary": "게시글 목록 조회",
        "parameters": [
          { "$ref": "#/components/parameters/start" },
          { "$ref": "#/components/parameters/end" },
          { "$ref": "#/components/parameters/sort" },
          { "$ref": "#/components/parameters/order" },
          { "name": "title_like", "in": "query", "schema": { "type": "string" }, "description": "제목 부분 일치" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["draft", "published", "archived"] } },
          { "name": "categoryId", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "게시글 목록",
            "headers": { "X-Total-Count": { "$ref": "#/components/headers/XTotalCount" } },
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Post" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["posts"],
        "summary": "게시글 생성",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PostInput" } } } },
        "responses": {
          "201": { "description": "생성됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Post" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/posts/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/idPath" }],
      "get": {
        "tags": ["posts"],
        "summary": "게시글 단건 조회",
        "responses": {
          "200": { "description": "게시글", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Post" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["posts"],
        "summary": "게시글 수정(부분)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PostInput" } } } },
        "responses": {
          "200": { "description": "수정됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Post" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["posts"],
        "summary": "게시글 삭제",
        "responses": {
          "200": { "description": "삭제 성공", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/categories": {
      "get": {
        "tags": ["categories"],
        "summary": "카테고리 목록 조회",
        "parameters": [
          { "$ref": "#/components/parameters/start" },
          { "$ref": "#/components/parameters/end" },
          { "$ref": "#/components/parameters/sort" },
          { "$ref": "#/components/parameters/order" },
          { "name": "name_like", "in": "query", "schema": { "type": "string" } },
          { "name": "parentId", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "카테고리 목록",
            "headers": { "X-Total-Count": { "$ref": "#/components/headers/XTotalCount" } },
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["categories"],
        "summary": "카테고리 생성",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryInput" } } } },
        "responses": {
          "201": { "description": "생성됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/categories/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/idPath" }],
      "get": {
        "tags": ["categories"],
        "summary": "카테고리 단건 조회",
        "responses": {
          "200": { "description": "카테고리", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["categories"],
        "summary": "카테고리 수정(부분)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CategoryInput" } } } },
        "responses": {
          "200": { "description": "수정됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Category" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["categories"],
        "summary": "카테고리 삭제",
        "responses": {
          "200": { "description": "삭제 성공", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/comments": {
      "get": {
        "tags": ["comments"],
        "summary": "댓글 목록 조회",
        "parameters": [
          { "$ref": "#/components/parameters/start" },
          { "$ref": "#/components/parameters/end" },
          { "$ref": "#/components/parameters/sort" },
          { "$ref": "#/components/parameters/order" },
          { "name": "postId", "in": "query", "schema": { "type": "integer" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["pending", "approved", "spam"] } },
          { "name": "authorName_like", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "댓글 목록",
            "headers": { "X-Total-Count": { "$ref": "#/components/headers/XTotalCount" } },
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["comments"],
        "summary": "댓글 생성",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommentInput" } } } },
        "responses": {
          "201": { "description": "생성됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/comments/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/idPath" }],
      "get": {
        "tags": ["comments"],
        "summary": "댓글 단건 조회",
        "responses": {
          "200": { "description": "댓글", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["comments"],
        "summary": "댓글 수정(부분)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommentInput" } } } },
        "responses": {
          "200": { "description": "수정됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["comments"],
        "summary": "댓글 삭제",
        "responses": {
          "200": { "description": "삭제 성공", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "관리자 로그인으로 발급된 JWT 액세스 토큰. `Authorization: Bearer <accessToken>`."
      }
    },
    "parameters": {
      "idPath": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "description": "리소스 ID" },
      "start": { "name": "_start", "in": "query", "schema": { "type": "integer", "minimum": 0 }, "description": "페이지네이션 시작 offset(포함)" },
      "end": { "name": "_end", "in": "query", "schema": { "type": "integer", "minimum": 0 }, "description": "페이지네이션 끝 offset(미포함)" },
      "sort": { "name": "_sort", "in": "query", "schema": { "type": "string" }, "description": "정렬 필드(콤마 구분)" },
      "order": { "name": "_order", "in": "query", "schema": { "type": "string" }, "description": "정렬 방향(asc|desc, 콤마 구분, _sort 와 1:1)" }
    },
    "headers": {
      "XTotalCount": { "description": "필터 적용 후 전체 레코드 수.", "schema": { "type": "integer" } }
    },
    "responses": {
      "BadRequest": { "description": "잘못된 요청", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "인증 실패(토큰 누락·만료·무효).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Forbidden": { "description": "권한 부족(인증됐으나 resource:action 권한 없음).", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "리소스를 찾을 수 없음", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Conflict": { "description": "충돌(예: slug/email 중복)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Post": {
        "type": "object",
        "required": ["id", "title", "slug", "content", "status", "tagIds", "tags", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "excerpt": { "type": "string" },
          "content": { "type": "string" },
          "status": { "type": "string", "enum": ["draft", "published", "archived"] },
          "categoryId": { "type": "integer", "nullable": true },
          "tagIds": { "type": "array", "items": { "type": "integer" } },
          "tags": { "type": "array", "items": { "$ref": "#/components/schemas/Tag" }, "readOnly": true },
          "publishedAt": { "type": "string", "format": "date-time", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
          "updatedAt": { "type": "string", "format": "date-time", "readOnly": true }
        }
      },
      "PostInput": {
        "type": "object",
        "required": ["title", "slug"],
        "properties": {
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "excerpt": { "type": "string" },
          "content": { "type": "string" },
          "status": { "type": "string", "enum": ["draft", "published", "archived"], "default": "draft" },
          "categoryId": { "type": "integer", "nullable": true },
          "tagIds": { "type": "array", "items": { "type": "integer" }, "default": [] }
        }
      },
      "Category": {
        "type": "object",
        "required": ["id", "name", "slug", "description", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "parentId": { "type": "integer", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
          "updatedAt": { "type": "string", "format": "date-time", "readOnly": true }
        }
      },
      "CategoryInput": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "parentId": { "type": "integer", "nullable": true }
        }
      },
      "Tag": {
        "type": "object",
        "required": ["id", "name", "slug", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
          "updatedAt": { "type": "string", "format": "date-time", "readOnly": true }
        }
      },
...
      "Comment": {
        "type": "object",
        "required": ["id", "postId", "authorName", "authorEmail", "content", "status", "createdAt", "updatedAt"],
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "postId": { "type": "integer" },
          "parentId": { "type": "integer", "nullable": true },
          "authorName": { "type": "string" },
          "authorEmail": { "type": "string", "format": "email" },
          "content": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "approved", "spam"] },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true },
          "updatedAt": { "type": "string", "format": "date-time", "readOnly": true }
        }
      },
      "CommentInput": {
        "type": "object",
        "required": ["postId", "authorName", "content"],
        "properties": {
          "postId": { "type": "integer" },
          "parentId": { "type": "integer", "nullable": true },
          "authorName": { "type": "string" },
          "authorEmail": { "type": "string", "format": "email" },
          "content": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "approved", "spam"], "default": "pending" }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["message"],
        "properties": { "message": { "type": "string" } }
      }
    }
  }
}